Use nginx user and move release into Makefile
Also change it to only release on master builds. To limit number of
images we have (I don't think we need branch builds? We can always push
it locally if we really need it).

And change default ingress port to 8080 so we don't need super user rights
to start up nginx. Needed so we can run as nginx in the Dockerfile.