Fix broken nginx.conf if no allow annotation
It was broken if allow annotation isn't specified. This change fixes
that. Basically, it was doing a split on an empty string, which produces
[]string{""}. This causes an `allow  ;` entry in nginx since it was
simply expanding the slice with `range`.

It also moves default allow into the controller, since everything that
gets an ingress entry should see the default allow as well. It's not
nginx specific.