Start Nginx only after first update is received
Prior to this change, there was a small window where the Nginx instance
used by feed-ingress could be active but without having loaded its
configuration. In this instance, the load balancer would be able to
reach it, but would receive a 404 response.

The solution is to only start Nginx after the first update has been
received and the configuration is ready - this means the load balancer
will only be able to reach it when it is suitably configured.

Fixes https://github.com/sky-uk/umc-core/issues/3586