Stop querying internal Nginx process state in test
Changed the tests in nginx_test.go so that they no longer rely on the
state of the nginx Process when making assertions.

Also fixed a problem in the TestFailsIfNginxDiesEarly test, due to the
deferral of Nginx startup to after the first update.

Renamed nginx.update() function to nginx.updateNginxConf(), to remove
possibility of confusion with the Update() function.