Start Nginx only after first update is received (#143)
* Start Nginx only after first update is received

Prior to this change, there was a small window where the Nginx instance
used by feed-ingress could be active but without having loaded its
configuration. In this instance, the load balancer would be able to
reach it, but would receive a 404 response.

The solution is to only start Nginx after the first update has been
received and the configuration is ready - this means the load balancer
will only be able to reach it when it is suitably configured.

Fixes https://github.com/sky-uk/umc-core/issues/3586

* Extra tests and healthcheck tweaks

Added extra tests to verify when the updater's Nginx process is
activated, also tweaked the healthcheck so the difference between an
Nginx that hasn't been started and one which has died is clear.

* Correcting golint failures

Fell foul of the golint rules (again). Fixed it up by creating a
non-exported function which returns a *nginxUpdater rather than a
controller.Updater.

* Stop querying internal Nginx process state in test

Changed the tests in nginx_test.go so that they no longer rely on the
state of the nginx Process when making assertions.

Also fixed a problem in the TestFailsIfNginxDiesEarly test, due to the
deferral of Nginx startup to after the first update.

Renamed nginx.update() function to nginx.updateNginxConf(), to remove
possibility of confusion with the Update() function.

* Addressing review comments

Specifically:
- Moved setting of InitialUpdateApplied flag back to after we've
  verified that Nginx has stayed up after starting.
- Added a better way of verifying that Nginx was started, by detecting
  the presence of a marker file in the temp directory.
- Made a couple of error assertions more specific to the nature of the
  error which arose.

* Fixed formatting failure.

Fell foul of Go's preferred formatting (again).

* Further changes following review.