Further changes
Added an example YAML file showing how to configure a deployment of Feed
DNS to point to load balancers identified by static hostnames.

Also removed the need to pass a region to the Route53 client session.
This isn't needed because Route53 isn't region-specific, and it looked a
bit odd to have to specify this when configuring Feed DNS to use static
hostnames rather than AWS load balancers.