Make ingress instance ip gorb specific
This is a concept that could also be used for both elb and alb
registration types. In fact we could extract the instance ip lookup to an
adapter that would decide whether to use it if provided or use the AWS
API to lookup the instance ip otherwise. Then refactor gorb, elb and alb
packages to make use of it.