Support for load balancers specified by hostname (#157)
* First pass at supporting internal and external load-balancer addresses as command line args for when we don't use ELB/ALB but still use R53.

* We're interested in A records plus CNAME records now.

* flatten r53 A and CNAME records into a common struct

* responded to PR comments
updated README and deprecated sky.uk/frontend-elb-scheme in favour of sky.uk/frontend-scheme to be LB agnostic

* responded to PR comments

* Checkpoint

* Refactoring, addressing review comments

Split all load balancer logic out into two separate adaptors, one for
AWS load balancers and one for load balancers referenced by static
hostnames.

* Handle updating of TTL on CNAME records

Fixed the issue identified by Seb, where we need to ensure the TTL value
on CNAME records is correct. (We may have decided to restart Feed-DNS
with a different TTL value inthe configuration.)

* Fixed formatting problems

* Addressing further review comments

* Further changes

Added an example YAML file showing how to configure a deployment of Feed
DNS to point to load balancers identified by static hostnames.

Also removed the need to pass a region to the Route53 client session.
This isn't needed because Route53 isn't region-specific, and it looked a
bit odd to have to specify this when configuring Feed DNS to use static
hostnames rather than AWS load balancers.

* Addressing the simple review comments from James

... before I do some more serious refactoring.

* Simplify logic around creating or updating records

Allows us to encapsulate the particular create/update behaviour which is
specific to each type of frontend.

* Push recognition of records down into adapters

Each frontend now has a recognise method which allows it to recognise
records on the behalf of the DNS updater. This removes some of the
frontend-specific logic from dns_updater.

* Fix missing validation case

Wasn't rejecting the case where an ALB was specified alongside a static
hostname.

* Refactor - moved adapters into their own package

This has the benefit of giving a cleaner interface between the adapters
and the core updater.

* Optional internal and external static hostnames

... although obviously at least one of them needs to be supplied.

Also updated the changelog and examples.

* Correct version number in CHANGELOG and examples

Since we're adding functionality in a backwards compatible manner, we
need to increase the minor version number, rather than the patch version
number.

* Method renaming in FrontendAdapter

Makes the purpose of the method clearer.