Changing gorb http client to pester
This is an in-place change of the http client to pester, a wrapper
for the built in golang client. No functionality has been added
with this commit, it is still set to use the default 5s timeout
with no retries.

Pester allows more granular control of the underlying http request
such as adding retries.