Adding ingress status updater tools
This is to add the underlying framework needed to add updater status
updaters which will allow us to set the status of an ingress.

Adding the ingress resource to the `IngressEntry` so that it can be used
at a later stage. Also including useful util functions which can be
reused by multiple status updaters.