Adding elb status updater
This is to add a new status updater for use with the elb's. It will
discover all of the elb's with the relevant labels (using existing
function as part of the elb updater). After this, it will iterate
through the ingress resources and set the relevent elb DNS name based on
the lb scheme label.