Add ingress loadbalancer status for elb and Merlin (#174)
* Adding ingress status updater tools

This is to add the underlying framework needed to add updater status
updaters which will allow us to set the status of an ingress.

Adding the ingress resource to the `IngressEntry` so that it can be used
at a later stage. Also including useful util functions which can be
reused by multiple status updaters.

* Adding elb status updater

This is to add a new status updater for use with the elb's. It will
discover all of the elb's with the relevant labels (using existing
function as part of the elb updater). After this, it will iterate
through the ingress resources and set the relevent elb DNS name based on
the lb scheme label.

* Adding Merlin status updater

This is to add a new status updater for use with Merlin. If enabled, the
user will need to specify both `internal-hostname` and `external-hostname`
as the updater uses these two values to match each ingress with the two
hostnames or ip addresses using the lb scheme label.

* Updating README and CHANGELOG

This is to add the changes that have been made to update the ingress
status to the README and create a new version in the CHANGELOG
`v1.10.0`.