Remove unnecessary config reload after start up
We start nginx after generating the config so we don't need to signal a
reload.  This was causing nginx to reload the config 5 minutes after
every startup which we know can be disruptive and is unnecessary.

Set the 'initial update' flag after starting nginx as this will ensure we
don't return true on the health endpoing before nginx is running.
Previously we set this flag just before starting nginx so it's
possible we were reporting healthy before the process was running.