Handle receiving 0 ingresses and prevent http 404s
Don't start Nginx or reload the config if Update is called with 0
ingresses.  Instead, return an error and let the controller handle it.
This fixes an issue where the k8s client returns 0 ingresses during
startup and Nginx then returns http 404s for all ingresses.

Fix for the sorting of ingresses when filtering for unique entries.
This fixes an issue with flappping ingresses when the same path is
defined for different services under the same host. This flapping
was causing Nginx to reload frequently which is disruptive.  We now
include serivce and port in the sorting.

Add a metric for Nginx config reloads.