Extra filtering for VTS zones
Misconfigured ingress resources have the ability to create invalid
VTS entries. These are long-lived and are not rectified againt the
running nginx config.

Because of this, it is possible that the metrics exported by feed
can flip between the valid and invalid VTS data causing erroneous
metrics to be exported.

This adds an extra filter to ensure that metrics are not generated
for misconfigured ingress resources.