validate paths passed to feed (#241)
k8s does not validate path entries that are part of ingresses, at least
not to the level of stringency that feed expects.  this can lead in some
cases to invalid nginx config being generated due to insufficient
santitzation.  the new approach is to reject paths which donʼt conform
strictly to the rfc 3986 definition of a path.

thanks @jsws for finding this bug