Add a readiness endpoint to feed. (#243)
The readiness endpoint calls down to a readiness method on the
controller, just as the healthcheck endpoint does.  The readiness
endpoint is implemented as a loop over all the controllerʼs owned
objects to poll them for readiness.

Most object types implement readiness by just returning the result of
their healthcheck.  The ALB and ELB types keep track of whether they have been registered with AWS, and return that status as their readiness.  The the NLB object type is where interesting things happen.  The AWS docs
call out a situation where it can take “a few minutes” for
healthchecking of an instance registered with a target group to begin.
During this interval, k8s should consider that replica “healthy” (itʼs
registered with the NLB successfully), but not “ready” (itʼs not
receiving traffic).  The distinction matters when doing rolling
deployments.  K8s wonʼt consider a new replica to be fully deployed (and move on with the deployment)
until itʼs ready.  And we donʼt want to treat feed as fully deployed
until it begins receiving traffic from the NLB target group (because we don't want to move on with deploying more replicas until the ones we have deployed are receiving traffic).

This situation has caused several production wobbles for us, where the
premature marking of feed pods as ready leads a rolling deployment to
complete too fast.  In the pessimistic case, we ended up with a brief
interval where there were no live replicas on the NLB at all (because
the old replicas had been torn down, but none of the new ones were out
of their NLB initial healthcheck delay yet).

The readiness logic for NLBs is so far implemented as a simple 5 minute
delay from initial NLB registration to the time when the NLB
controller (and by extension the entire feed process) will be marked as
“ready”.  In the cases weʼve seen, 5 minutes has been a sufficient
delay.  Thereʼs a possible better implementation, whereby we actually
observe whether weʼve received a healthcheck from NLB, and mark
ourselves as ready at the time when that occurs (which could be <5
minutes, leading to more responsive deployments, or could in theory be
\>5 minutes, giving us a safety net if the operational behavior of AWS
NLB changes).  However, thatʼs difficult to implement.  K8s and AWS
healthchecks arrive on the same endpoint.  The NLB controller has no
visibility of the request that is triggering a call to its health
method.  So the mooted implementation would require extensive
refactoring to be able to detect the "healthcheck from NLB" condition
specifically.

This implementation is good enough for now, and the better implementation
can be slotted in if/when it becomes a priority.

Having this implemented in feed itself allows us to avoid the need to
implement workarounds in the deployment yaml of feed in
kubernetes (replicating this implementation with
`initialReadinessDelay`, and not being forward-compatible with any
improvements).  It also moves the problem out of the realm of shadow
knowledge and into the open source implementation.