feat(feed): allow nginx to proxy WebSocket headers
Allows feed to support proxying WebSocket requests if explicitly enabled.

If enabled via the `--nginx-allow-websocket-upgrade` flag, the nginx configuration will be generated with the following block:

```
    # Support WebSocket upgrade, allow keepalive
    # Upgrade logic from http://nginx.org/en/docs/http/websocket.html
    map $http_upgrade $connection_upgrade {
        default upgrade;
        ''      '';
    }

    proxy_http_version 1.1;
    proxy_set_header Upgrade $http_upgrade;
    proxy_set_header Connection $connection_upgrade;
```

Which will set the `Connection` header to `upgrade` IF the `Upgrade` header is present, otherwise the `proxy_set_header Connection` will be `""` which does not forward, the same as existing feed behaviour.

If the flag is omitted, the nginx config will generate with:

```
    # Enable keepalive to backend.
    proxy_http_version 1.1;
    proxy_set_header Connection "";
```

as it previously did.