update all dependencies (#248)
the background to this change is that we are preparing to update feed to the networking/v1 Ingress API (from the extensions/v1beta1 API that was previously used, and is now deprecated/removed from recent versions of k8s).  in order to accomplish this, we need to update the k8s api libraries to more recent versions.  this is a good opportunity to rebaseline our dependencies and shed support for a deprecated module (merlin) within feed.  the next version of feed (incorporating the API change) will need to be a major version update, so we can batch these changes with the upcoming ones as well.

- update go version to 1.18
  - compiler 1.18 is required to build the newest `sigs.k8s.io/yaml` module
  - go.mod format 1.17 is required because of https://github.com/google/gnostic/issues/262
- use `go get -d -u ./...` to pull in the newest version of all dependencies
- delete the `merlin` module
  - its test suite relied on calling `proto.Clone` which does some reflect tricks with private fields that no longer work correctly
  - anyway, merlin is EOL so we don't need to keep supporting it
- update our usage of the k8s apis
  - pass a context to the functions that demand one
  - comply with the new update apis (passing a name of the field manager)
  - write some new deep-equal functions for a type (`LoadBalancerIngress`) that has fields (slices) in it now that are not comparable by simple `==`
- update the k8s test suite to match
  - use autogenerated mocks for the parts that were touched by this update, to make it easier to keep up with future updates
- update ginkgo to v2
  - this catches an issue in the gorb tests with a misplaced `Before/AfterSuite` invocation