fix(nginx.tmpl): port_in_redirect off
For locations with a trailing prefix like `/foo/`, nginx will redirect
any requests `/foo` to `/foo/`. By default, it redirects to the internal
listen port which likely will not match the port the external client
uses. Therefore, we should always preserve the external port the client
uses to access ingress.